Refactor: move Refresh Token generation logic to AccountManagement service
