+++ application/account-management/Core/Features/Authentication/Commands/RefreshAuthenticationTokens.cs
@@ -3,9 +3,9 @@ using System.Security.Claims;
 using JetBrains.Annotations;
 using Mapster;
 using Microsoft.AspNetCore.Http;
+using PlatformPlatform.AccountManagement.Features.Authentication.Domain.TokenGeneration;
 using PlatformPlatform.AccountManagement.Features.Users.Domain;
 using PlatformPlatform.SharedKernel.Authentication;
-using PlatformPlatform.SharedKernel.Authentication.TokenGeneration;
 using PlatformPlatform.SharedKernel.Cqrs;
 using PlatformPlatform.SharedKernel.Domain;
 using PlatformPlatform.SharedKernel.Telemetry;
