+++ application/account-management/Core/Features/EmailConfirmations/Commands/CompleteEmailConfirmation.cs
@@ -1,6 +1,6 @@
 using JetBrains.Annotations;
+using PlatformPlatform.AccountManagement.Features.Authentication.Domain;
 using PlatformPlatform.AccountManagement.Features.EmailConfirmations.Domain;
-using PlatformPlatform.SharedKernel.Authentication;
 using PlatformPlatform.SharedKernel.Cqrs;
 using PlatformPlatform.SharedKernel.Telemetry;
